+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-		Markdown Markup Parser (MDMP) - v1.1.0
+		Markdown Markup Parser (MDMP) - v1.1.1
 		Created by: Trevor W.
 
 		Github: https://github.com/trevor34/markdown-markup-parser/
@@ -19,6 +19,9 @@ All notable changes to this project will be documented in this file. The format 
 ## [Unreleased] - Upcoming Changes, Current Projects, and 'wish list' items
 - The ability to add selectors to anchors and images
 
+## [1.1.1] 2018-8-26
+## Fixed
+- Fixed bug in CDN files not parsing properly
 ## [1.1.0] 2018-8-23
 ## Added
 - new Regex detection for Start tags (`/!`)